Rework página, add admin date and url config, add and update new post UI
