+++ Models/HomeModel.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace GGJWeb.Models
@@ -9,7 +10,9 @@ namespace GGJWeb.Models
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
 
-        public DateTime? JamStart { get; set; }
+        public string? SignupLink { get; set; }
+
+        public DateTime JamStart { get; set; }
 
         public IEnumerable<Post>? posts { get; set; }
     }
