+++ Views/Home/Index.cshtml
@@ -1,10 +1,11 @@
-﻿@model IEnumerable<Post>
+﻿@model HomeModel
 
 @{
     ViewData["Title"] = "Home Page";
 }
 
-<header id="particles-js" class="homehead">
+<header class="homehead">
+    <div id="particles-js"></div>
     <div class="container position-relative px-4 px-lg-5">
         <div class="row gx-4 gx-lg-5 justify-content-center">
             <div class="col-md-10 col-lg-8 col-xl-7">
@@ -22,7 +23,7 @@
 <div class="container px-4 px-lg-5">
     <div class="row gx-4 gx-lg-5 justify-content-center">
         <div class="col-md-10 col-lg-8 col-xl-7">
-            @foreach (var post in Model)
+            @foreach (var post in Model.posts)
             {
                 <div class="post-preview">
                     <a asp-area="" asp-controller="Post" asp-action="Details" asp-route-id="@post.Id">
@@ -55,6 +56,8 @@
 
 <script>
     // Minus one because it's monthIndex so january is actually 0
-    const mainDate = new Date(2024, 2 - 1, 3, 17, 0, 0).getTime();
+    const mainDate = new Date("@Model.JamStart.ToString("yyyy-MM-ddTHH:mm:ss")").getTime();
 </script>
 
+<script src="~/js/particles.js" asp-append-version="true"></script>
+<script src="~/js/countdown.js" asp-append-version="true"></script>
