+++ wwwroot/js/particles.js
@@ -1,0 +1,5 @@
+﻿window.addEventListener('DOMContentLoaded', () => {
+    tsParticles.loadJSON('particles-js', 'particles.json').then(function (p) {
+        console.log('callback - particles.js config loaded');
+    });
+});
