+++ wwwroot/particles.json
@@ -1,7 +1,11 @@
 {
   "particles": {
+    "fullScreen": {
+      "enable": false,
+      "zIndex": 0
+    },
     "number": {
-      "value": 80,
+      "value": 60,
       "density": {
         "enable": true,
         "value_area": 800
@@ -27,7 +31,7 @@
     },
     "opacity": {
       "value": 0.5,
-      "random": false,
+      "random": true,
       "anim": {
         "enable": false,
         "speed": 1,
