+++ wwwroot/particles_admin.json
@@ -1,5 +1,10 @@
 {
+  "fullScreen": {
+    "enable": false,
+    "zIndex": 0
+  },
   "particles": {
+
     "number": {
       "value": 80,
       "density": {
@@ -26,8 +31,8 @@
       }
     },
     "opacity": {
-      "value": 0.5,
-      "random": false,
+      "value": 0.8,
+      "random": true,
       "anim": {
         "enable": false,
         "speed": 1,
@@ -49,12 +54,12 @@
       "enable": true,
       "distance": 150,
       "color": "#ffffff",
-      "opacity": 0.4,
-      "width": 2
+      "opacity": 0.3,
+      "width": 1.5
     },
     "move": {
       "enable": true,
-      "speed": 4,
+      "speed": 2,
       "direction": "top",
       "random": false,
       "straight": false,
